requirements.txt
aiocontextvars==0.2.2
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
contextvars==2.4
cryptography==2.9.2
Django==3.0.8
django-rest-knox==4.1.0
djangorestframework==3.11.0
dynaconf==3.0.0
gunicorn==20.0.4
hvac==0.10.4
idna==2.10
immutables==0.14
Jinja2==2.11.2
knox==0.1.0
loguru==0.5.1
MarkupSafe==1.1.1
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.1
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
